Added landscape layout

+++ app/src/main/java/com/example/friendsr/Friend.java
@@ -35,8 +35,7 @@ public class Friend implements Serializable {
         return rating;
     }
 
-    public void setRating(int rating) {
-
+    public void setRating(float rating) {
         this.rating = rating;
     }
 }
